refactor:💡 GetLessonProgressResponse 삭제 및 lessonResponse 공통화
